requirements.txt
Django==1.8.6
gunicorn==19.3.0
pep8==1.6.2
djangorestframework==3.3.2
django-rest-swagger==0.3.5
python-memcached==1.58
MySQL-python==1.2.5


# need this for JsonField in django
django-jsonfield==1.0.1

django-mysql==1.1.0

